- Fix build with libc++, dropping hardcoded -lstdc++ from the right Makefile. - Fix build with clang, casting a large constant to int in mntent_compat.c++. - Fix a noisy clang warning in BTree.h by parenthesizing the 2nd for() expr. - Fix a format string mismatch (%d/size_t -> %lu/unsigned long) in snprintf. - Regenerate patches, and normalize their filenames, with make makepatch. - Bump PORTREVISION.

+++ devel/fam/files/mntent_compat.c++
@@ -61,7 +61,7 @@ struct {
 	{ MNT_SYNCHRONOUS,	"sync"},
 	{ MNT_UNION,		"union"},
 	{ MNT_NOCLUSTERR,	"noclusterr"},
-	{ MNT_NOCLUSTERW,	"noclusterw"},
+	{ static_cast<int>(MNT_NOCLUSTERW),	"noclusterw"},
 	{ MNT_SUIDDIR,		"suiddir"},
 #ifdef MNT_SNAPSHOT
 	{ MNT_SNAPSHOT,		"snapshot"},
